Update logbook.md
So yesterday I did the memory display for the viewer.

Today I worked more on the wrapper, very close to MVP on that one now.
Only need some more functions and some testing and hopefully it does what it's supposed to do.

Also updated the dates on this logbook, how on earth did I not see the year was wrong on most of the entries?

Took 9 minutes