Model options
Added the option to include model parameters that carry over to the GUI when loaded. This makes it possible to load a model or assembly, change the options in the GUI and the model will rebuild.

Cleaned up and removed "old" files. The TK is out, no turning back from PySide6.

Removed some of the modules like Titan, Bombe and Enigma. Those were based upon old ideas, most of which is now redundant with the new concept.

Took 2 hours 7 minutes