Model semantics
Changed model behavior to return a list of shapes.
The viewer converts that into mesh and displays it.
This should be a step on the way to have BOM data and drawing generated from the model.

Reworked several parts of the wrapper, fixed some issues with the rotation. Added more tests.

Took 2 hours 17 minutes