Linux Wrapper
It works! At least on my machine, feedback if there is still issues would be great.

Run the setup script to compile the rest of the files needed for Linux.
I could add them, but since they take up much more space than Windows do, and the fact that it's easier to get going compiling on Linux I won't if I don't have to.