Instances
From 125,000 cubes in 88 seconds and 2.27 GB file to 1,000,000 in 104 seconds and a 621 MB file.

Thought a 1 million cube step file is pointless, it shows the ability to output a massive assembly with ease.

And this was done with a Python single thread worker.

A large assembly made up of more complex shapes will be able to be multi threaded.

Took 4 hours 47 minutes